Module Byterntm

Parser for RNTM in bytecode executables. Parses both the RNTM section and the shebang launcher produced by Bytelink.

type search_method =

| Disable of string
(*
Check fixed location only
*)

| Fallback of string
(*
Check given location first then fallback to searching for the interpreter
*)

| Enable
(*
Always search for the interpreter
*)

Search methods used by a tendered bytecode image to find a runtime.

val read_runtime : Bytesections.section_table -> in_channel -> (string * Misc.RuntimeID.t option * search_method) option

Returns the runtime used by this tendered/standalone image. If the runtime used cannot be parsed, or the image was linked using -without-runtime, then None is returned.