Module Env.Unscoped

alert dangerous It is unsafe to use the common [Path.same] function in contexts where [Unscoped.with_pairs] has been used to enrich the environment with unscoped equalities; [Path.equiv] must be used instead. See [ctype.ml] for an example of careful usage of [Unscoped], by locally shadowing [Path] to hide [Path.same].

val with_pairs : (Ident.Unscoped.t * Ident.Unscoped.t) list -> t -> t

val get_pairs : t -> (Ident.Unscoped.t * Ident.Unscoped.t) list

val path_equiv : t -> Path.t -> Path.t -> bool