Module Load_path.Dir

type t

Represent one directory in the load path.

val create : string -> t

val path : t -> string

val files : t -> string list

All the files in that directory. This doesn't include files in sub-directories of this directory.

val find : t -> string -> string option

find dir fn returns the full path to fn in dir.

val find_uncap : t -> string -> string option

As find, but search also for uncapitalized name, i.e. if name is Foo.ml, either /path/Foo.ml or /path/foo.ml may be returned.