Module Printtyp.Conflicts

The Conflicts module keeps track of conflicts arising when attributing names to identifiers and provides functions that can print explanations for these conflict in error messages

val exists : unit -> bool

exists() returns true if the current naming context renamed an identifier to avoid a name collision

type explanation = {

kind : Shape.Sig_component_kind.t;

name : string;

root_name : string;

location : Location.t;

}

val list_explanations : unit -> explanation list

list_explanations() return the list of conflict explanations collected up to this point, and reset the list of collected explanations

val print_located_explanations : Format.formatter -> explanation list -> unit

val print_explanations : Format.formatter -> unit

Print all conflict explanations collected up to this point

val reset : unit -> unit